+++ MoneyIO/app/src/main/java/io/money/moneyio/model/database/DatabaseHelperFirebase.java
@@ -218,7 +218,7 @@ public class DatabaseHelperFirebase {
             MoneyFlow t = dataSnapshot.getValue(MoneyFlow.class);
             data.add(t);
             if(SystemClock.elapsedRealtime() - elapsedTime > 3000){
-                Utilities.notifyFriend(context, t.getType(), Float.toString(t.getSum()));
+                Utilities.notifyFriend(context, t.getType(), Double.toString(t.getSum()));
             }
         }
